requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-google-genai==1.0.1
streamlit==1.32.0
chromadb==0.4.24
pypdf==4.1.0
python-dotenv==1.0.1
google-generativeai==0.4.1
sentence-transformers==2.7.0
